One of the subtests in TestGrassAlgorithmsRasterTest part 2 fails.
It reads a yaml file that specifies algorithms, inputs for these, and expected output.

In the yaml file, its number 15 (counting from ), with name "GRASS7 r.latlong", that fails.

But when it fails, subsequently there is a segfault. So something is wrong besides the fact that the result does not match any of the accepted results.

So to demonstrate this, and perhaps even resolve it, I am building a simpler test case, that demonstrates this behavior.

I call the new file segfault.py. I created the first version by merging TestGrassAlgorithmsRasterTest (part 2) and AlgorithmsTestBase.

I have now seen that it still fails and causes segfault. I will now continue making the demo simpler.